type de projet et quelque controller ok

+++ CrowdfundingSout/src/main/java/com/CrowdfundingSoutenance/CrowdfundingSout/ServicesImplemetion/RealisationServImplementa.java
@@ -16,18 +16,32 @@ public class RealisationServImplementa implements RealisationInterface {
     RealisationRepository realisationRepository;
 
     @Override
-    public Realisation createRealisation(Realisation realisation) {
-        return realisationRepository.save(realisation);
+    public String createRealisation(Realisation realisation) {
+        if (realisation != null) {
+            realisationRepository.save(realisation);
+            return "Realisation ajouter ";
+        }
+        return "Erreur";
     }
 
     @Override
-    public Realisation UpdateRealisationById(Long id, Realisation realisation) {
-        return null;
+    public String UpdateRealisationById(Long IdRealisation, Realisation realisation) {
+        Realisation existingRealisation = realisationRepository.findById(IdRealisation).orElse(null);
+        if (existingRealisation != null && realisation != null) {
+            existingRealisation.setNomRealisation(realisation.getNomRealisation());
+            existingRealisation.setMontantRealisation(realisation.getMontantRealisation());
+            existingRealisation.setImageRealisation(realisation.getImageRealisation());
+            existingRealisation.setDaterealisation(realisation.getDaterealisation());
+            realisationRepository.save(existingRealisation);
+
+            return "Réalisation modifier avec succès";
+        }
+        return "Realisation non modifier";
     }
 
     @Override
-    public Realisation getRealisationById(Long id) {
-        return realisationRepository.findById(id).orElse(null);
+    public Realisation getRealisationById(Long IdRealisation) {
+        return realisationRepository.findById(IdRealisation).orElse(null);
     }
 
     @Override
@@ -36,12 +50,18 @@ public class RealisationServImplementa implements RealisationInterface {
     }
 
     @Override
-    public List<Realisation> getRealisationByName(String name) {
-        return realisationRepository.findByNomRealisation(name);
+    public List<Realisation> getRealisationByName(String nomRealisation) {
+        if(nomRealisation != null && !nomRealisation.isEmpty()){
+            return realisationRepository.findByNomRealisation(nomRealisation);
+        }
+        return null;
     }
 
     @Override
     public void deleteRealisation(Long id) {
-        realisationRepository.deleteById(id);
+        Realisation existingRealisation = realisationRepository.findById(id).orElse(null);
+        if(existingRealisation != null){
+            realisationRepository.deleteById(id);
+        }
     }
 }
